requirements.txt
torch~=1.11.0
opencv-python~=4.5.5.64
numpy~=1.22.3
python-dotenv~=0.20.0
torchvision~=0.12.0
PyJWT~=2.3.0
Flask~=2.1.0
Werkzeug~=2.1.0
gunicorn~=20.1.0
